+++ d/phobos2/std/outbuffer.d
@@ -69,7 +69,7 @@ class OutBuffer
      */
 
 
-    void reserve(uint nbytes)
+    void reserve(size_t nbytes)
 	in
 	{
 	    assert(offset + nbytes >= offset);
@@ -328,7 +328,7 @@ class OutBuffer
      * all data past index.
      */
 
-    void spread(uint index, uint nbytes)
+    void spread(size_t index, size_t nbytes)
 	in
 	{
 	    assert(index <= offset);
